fix: インストールスクリプトを修正

+++ install.sh
@@ -2,49 +2,33 @@
 set -euox pipefail
 cd "$(dirname "$0")"
 
-URL="https://github.com/MIDORIBIN/offline-vscode-extensions/tarball/main"
-wget --no-check-certificate -O - ${URL} | tar zxvf -
-cd MIDORIBIN-offline-vscode-extensions-*
+if [ ${#} != 1 ]; then
+    echo "invalid args length: ${#}"
+    exit 1
+fi
+
+if [ ${1} != "host" ] && [ ${1} != "container" ]; then
+    echo "invalid args: ${1}"
+    exit 1
+fi
 
 
-COMMON_VSIX_DIR="vsix/common"
-HOST_VSIX_DIR="vsix/host"
-CONTAINER_VSIX_DIR="vsix/container"
-
-# host
 OUT="${HOME}/.vscode-server/extensions"
+COMMON_VSIX_DIR="vsix/common"
+SPECIFIC_VSIX_DIR="vsix/${1}"
+
 rm -rf ${OUT}
 mkdir -p ${OUT}
 
-for vsix in ${COMMON_VSIX_DIR}/*.tar.gz; do
-    NAME=${vsix##*/}
-    NAME=${NAME%.*}
-    tar xvzf ${vsix} -C ${OUT}
+VISXS=(${COMMON_VSIX_DIR}/* ${SPECIFIC_VSIX_DIR}/*)
+
+
+function untar() {
+    local out=${1}
+    local vsix=${2}
+    tar xvzf ${vsix} -C ${OUT} 
+}
+
+for vsix in "${VISXS[@]}"; do
+    untar ${OUT} ${vsix}
 done
-
-for vsix in ${HOST_VSIX_DIR}/*.tar.gz; do
-    NAME=${vsix##*/}
-    NAME=${NAME%.*}
-    tar xvzf ${vsix} -C ${OUT}
-done
-
-
-# container
-OUT="${HOME}/.vscode-server/container-extensions"
-rm -rf ${OUT}
-mkdir -p ${OUT}
-
-for vsix in ${COMMON_VSIX_DIR}/*.tar.gz; do
-    NAME=${vsix##*/}
-    NAME=${NAME%.*}
-    tar xvzf ${vsix} -C ${OUT}
-done
-
-for vsix in ${CONTAINER_VSIX_DIR}/*.tar.gz; do
-    NAME=${vsix##*/}
-    NAME=${NAME%.*}
-    tar xvzf ${vsix} -C ${OUT}
-done
-
-cd ../
-rm -r MIDORIBIN-offline-vscode-extensions-*
